chore: create json schema file for find source report, and a valdation test for it
